+++ tests/conftest.py
@@ -7,14 +7,15 @@ from collections import defaultdict
 from pathlib import Path
 from subprocess import run
 
+import pytest
 import vdf
-from protontricks.cli.main import main
-from protontricks.cli.launch import main as launch_main
-from protontricks.cli.desktop_install import main as desktop_install_main
+
+from protontricks.cli.desktop_install import \
+    cli as desktop_install_cli_entrypoint
+from protontricks.cli.launch import cli as launch_cli_entrypoint
+from protontricks.cli.main import cli as main_cli_entrypoint
 from protontricks.steam import (APPINFO_STRUCT_HEADER, APPINFO_STRUCT_SECTION,
                                 SteamApp, get_appid_from_shortcut)
-
-import pytest
 
 
 @pytest.fixture(scope="function", autouse=True)
@@ -607,6 +608,10 @@ def zenity(monkeypatch):
         "protontricks.gui.run",
         mock_subprocess_run
     )
+    monkeypatch.setattr(
+        "protontricks.cli.util.run",
+        mock_subprocess_run
+    )
 
     yield mock_zenity
 
@@ -686,7 +691,7 @@ def cli(monkeypatch, capsys):
     Run `protontricks` with the given arguments and environment variables,
     and return the output
     """
-    return _run_cli(monkeypatch, capsys, main)
+    return _run_cli(monkeypatch, capsys, main_cli_entrypoint)
 
 
 @pytest.fixture(scope="function")
@@ -695,7 +700,7 @@ def launch_cli(monkeypatch, capsys):
     Run `protontricks-launch` with the given arguments and environment
     variables, and return the output
     """
-    return _run_cli(monkeypatch, capsys, launch_main)
+    return _run_cli(monkeypatch, capsys, launch_cli_entrypoint)
 
 
 @pytest.fixture(scope="function")
@@ -704,4 +709,4 @@ def desktop_install_cli(monkeypatch, capsys):
     Run `protontricks-desktop-install` with the given arguments and environment
     variables, and return the output
     """
-    return _run_cli(monkeypatch, capsys, desktop_install_main)
+    return _run_cli(monkeypatch, capsys, desktop_install_cli_entrypoint)