+++ CHANGELOG.md
@@ -9,6 +9,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
  - Add `protontricks-launch` script to launch Windows executables using Proton app specific Wine prefixes
  - Add desktop integration for Windows executables, which can now be launched using Protontricks
  - Add `protontricks-desktop-install` to install desktop integration for the local user. This is only necessary if the installation method doesn't do this automatically.
+ - Add error dialog for displaying error information when Protontricks has been launched from desktop and no user-visible terminal is available.
 
 ### Changed
  - Improve GUI dialog. The prompt to select the Steam app now uses zenity's "list" dialog, which supports scrolling and search.